Функция isConnected вынесена в common. Удалены лишние константы в SearchFragment
