+++ app/src/main/java/ru/practicum/android/diploma/util/CheckInternet.kt
@@ -1,0 +1,18 @@
+package ru.practicum.android.diploma.util
+
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+
+fun isConnected(context: Context): Boolean {
+    val connectivityManager = context.getSystemService(
+        Context.CONNECTIVITY_SERVICE
+    ) as ConnectivityManager
+    val capabilities = connectivityManager.getNetworkCapabilities(connectivityManager.activeNetwork)
+
+    return capabilities != null && (
+        capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR)
+            || capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI)
+            || capabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET)
+        )
+}
